pyproject.toml
[tool.poetry]
name = "twinplayerbot"
version = "0.1.0"
description = ""
authors = ["Sergey Popov <y.sergey.popov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.47.0"
python-telegram-bot = "^21.6"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
